+++ playhouse/postgres_ext.py
@@ -726,11 +726,8 @@ class PostgresqlExtDatabase(PostgresqlDatabase):
     def execute_sql(self, sql, params=None, named_cursor=None):
         if not named_cursor:
             return super(PostgresqlExtDatabase, self).execute_sql(sql, params)
-        self._log_query(sql, params)
-        with __exception_wrapper__:
-            cursor = self.cursor(named_cursor=named_cursor)
-            cursor.execute(sql, params or ())
-        return cursor
+        return self._execute_cursor(sql, params,
+                                    {'named_cursor': named_cursor})
 
 
 class PooledPostgresqlExtDatabase(_PooledPostgresqlDatabase, PostgresqlExtDatabase):